+++ tests/test_replies.py
@@ -6,6 +6,7 @@ from nose.tools import assert_raises
 
 from werobot.parser import parse_user_msg
 from werobot.replies import WeChatReply, TextReply, ImageReply, MusicReply
+from werobot.replies import VoiceReply
 from werobot.replies import Article, ArticlesReply
 from werobot.replies import TransferCustomerServiceReply, SuccessReply
 from werobot.utils import to_binary, to_text
@@ -63,6 +64,25 @@ def test_image_reply():
     <Image>
     <MediaId><![CDATA[fdasfdasfasd]]></MediaId>
     </Image>
+    </xml>""".format(time=t).strip()
+
+
+def test_voice_reply():
+    t = int(time.time())
+    reply = VoiceReply(
+        target='tgu',
+        source='su',
+        media_id="fdasfdasfasd", time=t
+    )
+    reply.render().strip() == """
+    <xml>
+    <ToUserName><![CDATA[tgu]]></ToUserName>
+    <FromUserName><![CDATA[su]]></FromUserName>
+    <CreateTime>{time}</CreateTime>
+    <MsgType><![CDATA[voice]]></MsgType>
+    <Voice>
+    <MediaId><![CDATA[fdasfdasfasd]]></MediaId>
+    </Voice>
     </xml>""".format(time=t).strip()